@LucasWilkinson 2025-02-25
> @LucasWilkinson , we upstreamed our change to groupwise scaling kernels. there are some conflicts in this PR that needs to be solved.
> 
> Our change is mainly:
> 
> ```
> Extend groupwise scaling gemm to support both M dimension and N dimension groupwise scaling in FP8 GEMM.
> In examples/67_hopper_fp8_warp_specialized_gemm_with_blockwise_scaling/67_hopper_fp8_warp_specialized_gemm_with_groupwise_scaling.cu, two parameters ScaleGranularityM and ScaleGranularityNcontrol the scaling mode:
> 
> 
> ScaleGranularityM == 128 and ScaleGranularityN == 128 --> 2Dx2D scaling (block-wise scaling, same as 67_hopper_fp8_warp_specialized_gemm_with_blockwise_scaling.cu , 2Dx2D refers to the shape of the scaling factor)
> 
> ScaleGranularityM == 1 and ScaleGranularityN == 128 --> 1Dx2D scaling
> 
> ScaleGranularityM == 128 and ScaleGranularityN == 1 --> 2Dx1D scaling
> 
> ScaleGranularityM == 1 and ScaleGranularityN == 1 --> 1Dx1D scaling
> ```

apologies for the delay the PR has been updated, currently I am still vectorizing the loads of B scales along N (like `main`) but it might actually makes sense to not do this to enable transposing A and B (since we currently have partial tiles along M this would mean partial tiles along N)