@qijiaxing 2025-03-06
Hi @LucasWilkinson ,
if I want to pass pytoch tensors to do a 1D2D GEMM, should be the tensor and scale tensor shape like this?
```
A: [M, K]
B: [N, K]
A scale: [K/128, M]
B scale: [K/128, N/128]
```